Move /friends route into a friends blueprint and fix bidirectional friend query
