+++ templates/base.html
@@ -32,7 +32,7 @@
 
         <div class="nav-center">
           <a href="{{ url_for('index') }}">Home</a>
-          <a href="{{ url_for('friends') }}">Friends</a>
+          <a href="{{ url_for('friends.friends') }}">Friends</a>
           <a href="{{ url_for('add_event') }}">Add Event</a>
         </div>
       </div>
